pull out shared logic in member and bill queries/tables
also revise some api & table logic to use more generic fields vs specific ones. This makes it easier to share logic
