+++ src/php/api/routes/offices/class.offices.by.bioguide.id.php
@@ -3,11 +3,11 @@
 namespace API {
     class OfficesByBioguideId extends OfficesRoute {
 
-        public static function parameters() { return ["id"]; }
+        public static function parameters() { return ["bioguideId"]; }
         
         public static function fetchResult() {
-            $id = Parameters::get("id");
-            return \AuditCongress\MemberOffices::getByBioguideId($id);
+            $bioguideId = Parameters::get("bioguideId");
+            return \AuditCongress\MemberOffices::getByBioguideId($bioguideId);
         }
     }
 }
