+++ src/php/api/routes/offices/class.offices.by.office.id.php
@@ -3,11 +3,11 @@
 namespace API {
     class OfficesByOfficeId extends OfficesRoute {
 
-        public static function parameters() { return ["officeid"]; }
+        public static function parameters() { return ["id"]; }
         
         public static function fetchResult() {
-            $officeid = Parameters::get("officeid");
-            return \AuditCongress\MemberOffices::getByOfficeId($officeid);
+            $id = Parameters::get("id");
+            return \AuditCongress\MemberOffices::getById($id);
         }
     }
 }
