+++ src/php/audit.congress/objects/bills/querys/class.bill.titles.query.php
@@ -4,7 +4,7 @@ namespace AuditCongress {
 
     class BillTitlesQuery extends AuditCongressQuery {
 
-        use BillsGetByIdQuery, BillsGetByBillIdQuery, BillsGetWithFilterQuery;
+        use GetByIdQuery, BillsGetByBillIdQuery, BillsGetWithFilterQuery;
 
         public function __construct() {
             parent::__construct("BillTitles");
