+++ src/php/audit.congress/objects/bills/tables/class.bill.titles.php
@@ -4,7 +4,7 @@ namespace AuditCongress {
 
     class BillTitles extends BillTable {
 
-        use BillsGetById, BillsGetByBillId;
+        use GetById, BillsGetByBillId;
 
         private function __construct() {
             parent::__construct("BillTitles", "\AuditCongress\BillTitlesQuery");
