+++ src/php/audit.congress/objects/members/rows/class.member.offices.row.php
@@ -4,18 +4,18 @@ namespace AuditCongress {
 
     class MemberOfficesRow extends \MySqlConnector\SqlRow {
         public
-            $bioguideId,$officeId,
+            $bioguideId,$id,
             $address,$suite,$building,$city,$state,$zip,
             $latitude,$longitude,$phone,$fax;
 
         public function getColumns() {
-            return ["bioguideId","officeId","address","suite",
+            return ["bioguideId","id","address","suite",
                     "building","city","state","zip","latitude",
                     "longitude","phone","fax"];
         }
 
         public function getValues() {
-            return [$this->bioguideId,$this->officeId,$this->address,
+            return [$this->bioguideId,$this->id,$this->address,
             $this->suite,$this->building,$this->city,$this->state,
             $this->zip,$this->latitude,$this->longitude,$this->phone,$this->fax];
         }
