+++ src/php/audit.congress/objects/members/tables/class.member.offices.php
@@ -3,7 +3,7 @@
 namespace AuditCongress {
 
     class MemberOffices extends MemberTable {
-
+        use GetByBioguideId, GetById;
         private function __construct() {
             parent::__construct("MemberOffices", "\AuditCongress\MemberOfficesQuery");
         }
@@ -17,17 +17,6 @@ namespace AuditCongress {
 
         protected static function parseResult($resultRows) {
             return MemberOfficesRow::rowsToObjects($resultRows);
-        }
-
-        public static function getByBioguideId($bioguideId) {
-            $offices = MemberOfficesQuery::getByBioguideId($bioguideId);
-            return self::parseResult($offices);
-        }
-
-        public static function getByOfficeId($officeId) {
-            self::enforceCache();
-            $offices = MemberOfficesQuery::getByOfficeId($officeId);
-            return self::parseResult($offices);
         }
     }
 }
