+++ src/php/audit.congress/objects/members/tables/class.member.socials.php
@@ -3,7 +3,8 @@
 namespace AuditCongress {
 
     class MemberSocials extends MemberTable {
-        
+        use GetByBioguideId;
+
         private function __construct() {
             parent::__construct("MemberSocials", "\AuditCongress\MemberSocialsQuery");
         }
@@ -17,12 +18,6 @@ namespace AuditCongress {
 
         protected static function parseResult($resultRows) {
             return MemberSocialsRow::rowsToObjects($resultRows);
-        }
-
-        public static function getByBioguideId($bioguideId) {
-            self::enforceCache();
-            $socials = MemberSocialsQuery::getByBioguideId($bioguideId);
-            return self::returnFirst(self::parseResult($socials));
         }
     }
 }
